pyproject.toml


[tool.hatch.envs.default]
dependencies = [
    "pandas>=1.3.0",
    "sdv>=1.0.0",
    "awswrangler>=3.0.0",
    "boto3>=1.26.0",
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "nbconvert",
    "jupyter",
    "jupyterlab",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
    "sphinx-autodoc-typehints>=1.12.0",
    "linkify-it-py>=2.0.0"
]

[tool.hatch.envs.default.scripts]
docs-build = "cd docs && make html"
docs-clean = "cd docs && make clean"
docs-serve = "python -m http.server 8000 --directory docs/_build/html"



[tool.hatch.envs.light]
skip-install = false
dependencies = [
    "pandas>=1.3.0",
    "pyyaml",
    "pytest>=7.0",
    "flake8>=4.0",
    "artifact-store @ git+https://github.com/eisenhauerIO/utils-artifacts-store.git@bc3a7d6e02266ddb24348279f75bf314de16a21a"
]
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "online-retail-simulator"
version = "0.1.0"
description = "A lightweight Python package for simulating realistic retail data for experimentation and causal inference"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Eisenhauer IO"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pandas>=1.3.0",
    "pyyaml",
    "artifact-store @ git+https://github.com/eisenhauerIO/utils-artifacts-store.git@bc3a7d6e02266ddb24348279f75bf314de16a21a",
]


[project.optional-dependencies]
synthesizer = [
    "sdv>=1.0.0"
]
cloud = [
    "artifact-store[cloud] @ git+https://github.com/eisenhauerIO/utils-artifacts-store.git@bc3a7d6e02266ddb24348279f75bf314de16a21a",
]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "nbconvert",
    "jupyter",
    "jupyterlab",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
    "sphinx-autodoc-typehints>=1.12.0",
    "linkify-it-py>=2.0.0"
]

light = [
    "pandas>=1.3.0",
    "pyyaml",
    "flake8>=4.0"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.setuptools.packages.find]
where = ["."]
include = ["online_retail_simulator*"]

[tool.setuptools.package-data]
online_retail_simulator = ["config_defaults.yaml"]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
